Make tests actually detect errors/assertions in handlers
Previously, some tests were promise-chaining handlers containing
testthat assertions, and then calling wait_for_it(). This would
cause those assertions to be printed but not treated as test
failures, because wait_for_it() doesn't error on unhandled
promise errors. With this change, wait_for_it() takes a promise
and will raise an error if the promise rejects.

The bigger part of the commit is revisiting all the existing
uses of wait_for_it() and passing in promises if appropriate.